deploy/requirements.txt
streamlit>=1.28.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scipy>=1.11.0
requests>=2.31.0
python-dotenv>=1.0.0
twilio>=8.10.0
